Changes to Passport and State JavaScript files.
Changed state.js to make it static so that it can be accessed anywhere.
Changed sidebar.js to make use of state.js being static.
Changed passport.js to include the render function.
Changed canvas.js to remove the state variable.
Changed canvas.game.js to remove the line that rendered the passport.