Passport can now be moved.
Modified state.js to include a timer that can be used to update the
passport and added in a tempory inclusion of passport-move to test move
code.
Created passport-move.js - test playground for movement and boundary
code, will be put into passport.js at some point.
Modified input to better check if the mouse is in bounds and added code
to help with object movement and boundaries.
Modified canvas.menu.js to return a size.
Modified canvas.game.js to return a size and render the actual passport
+ test passport.
Modified canvas.js to be able to return the active canvas.
Modified references.php to include the test passport.